refactor: move IndieHoy parser to musicDiscoveryController and update error handling
