+++ controllers/musicDiscoveryController.js
@@ -1,4 +1,3 @@
-import { indieHoyParser } from "./musicArticleController.js";
 import { discover } from "../services/musicDiscoveryService.js";
 import {
   searchSong,
@@ -7,6 +6,77 @@ import {
   YoutubeMusicAuthError,
   YoutubeMusicRateLimitError,
 } from "../services/youtubeMusicService.js";
+
+// --- IndieHoy Parser ----------------------------------------------------------
+// Site-specific knowledge (selectors, regex) lives here, not in services.
+// When a new music source is added, it gets its own parser block in its
+// own controller — each controller owns its source's DOM shape.
+
+const SELECTORS = {
+  ENTRY_CONTENT: ".post-content-wrap .entry-content",
+  HEADINGS: "h2.wp-block-heading",
+};
+
+/**
+ * Extracts artist/song pairs from an IndieHoy article's DOM.
+ *
+ * Strategy:
+ * 1. Scopes to .post-content-wrap .entry-content to avoid matching
+ *    headings outside the article body.
+ * 2. Iterates h2.wp-block-heading elements.
+ * 3. Matches each heading against two regexes (text quotes then HTML <em>)
+ *    to extract artist and song. Non-matching headings (Conclusion, etc.)
+ *    are skipped with a console.warn for debugging.
+ *
+ * @param {import("cheerio").CheerioAPI} $ - Cheerio instance loaded with the article HTML.
+ * @returns {{ artist: string, song: string }[]}
+ */
+export const indieHoyParser = ($) => {
+  const items = [];
+  const scope = $(SELECTORS.ENTRY_CONTENT);
+
+  scope.find(SELECTORS.HEADINGS).each((_, el) => {
+    const $h2 = $(el);
+    const text = $h2.text().trim();
+    // .html() keeps entities like &nbsp; as-is — strip them so the
+    // $ anchor in the HTML regex isn't blocked by trailing non-breaking spaces.
+    const html = ($h2.html() || "").replace(/&nbsp;/g, "").trim();
+
+    // Try text regex first: matches "Artist - "Song"" (handles both
+    // straight quotes U+0022 and smart/curly quotes U+201C/U+201D).
+    const textMatch = text.match(/^(.+?)\s+-\s+["\u201c](.+?)["\u201d]$/);
+    if (textMatch) {
+      items.push(buildItem(textMatch[1], textMatch[2]));
+      return;
+    }
+
+    // Try HTML regex: matches "Artist - <em>Song</em>"
+    const htmlMatch = html.match(/^(.+?)\s+-\s+<em>(.+?)<\/em>$/i);
+    if (htmlMatch) {
+      items.push(buildItem(htmlMatch[1], htmlMatch[2]));
+      return;
+    }
+
+    // No match — skip non-music headings (Conclusion, etc.)
+    console.warn(`[indieHoyParser] Skipping heading with no artist/song pattern: "${text}"`);
+  });
+
+  return items;
+};
+
+/**
+ * Builds a single parsed item from a heading element.
+ *
+ * @param {string} artist - Artist name extracted via regex.
+ * @param {string} song - Song title extracted via regex.
+ * @returns {{ artist: string, song: string }}
+ */
+const buildItem = (artist, song) => {
+  return {
+    artist: artist.trim(),
+    song: song.trim(),
+  };
+};
 
 // --- IndieHoy Source Config ---------------------------------------------------
 
@@ -115,21 +185,16 @@ export const getMusicDiscovery = async (req, res) => {
       // Step 4 — add found tracks to the playlist
       const videoIds = tracksAdded.map((t) => t.videoId);
       const addResult = await addTracksToPlaylist(created.playlistId, videoIds);
-
-      // Reconcile: move skipped tracks from tracksAdded to tracksNotFound
-      if (addResult.skipped > 0) {
-        const actuallyAdded = tracksAdded.slice(0, addResult.trackCount);
-        const skippedTracks = tracksAdded.slice(addResult.trackCount);
-        skippedTracks.forEach((t) =>
-          tracksNotFound.push({
-            title: t.title,
-            artist: t.artist,
-            reason: "Skipped during playlist add (unavailable)",
-          })
-        );
-        tracksAdded.length = 0;
-        tracksAdded.push(...actuallyAdded);
-      }
+      // addResult tracks which videoIds were successfully added (including
+      // 409 duplicates) and how many were skipped (unavailable). We do NOT
+      // try to reclassify tracksAdded items — the index-based approach would
+      // be wrong when skips are interleaved with successes. Instead we report
+      // the add result alongside the search results so the caller can reconcile.
+      playlist.addToPlaylist = {
+        attempted: videoIds.length,
+        added: addResult.trackCount,
+        skipped: addResult.skipped,
+      };
     }
 
     // Step 5 — build and return report (strip videoId from response)
